Enable main window dragging
This commit adds functionality to the main window of the application to allow it to be dragged around by the user. The user can now click and hold the left mouse button anywhere on the window and move the mouse to drag the window to a new location on the screen.