Added favorite feature with indicators.
In this commit, I added the favorite feature that allows the user to mark games as favorites. 
I added a golden border around the game to indicate that it is marked as a favorite. 
The user can toggle the favorite status of a selected game by clicking on the star icon.