Updated README file to add user introduction on how to use the program
In this commit, I updated the README file to provide the user with a simple introduction on how to use the program. 
I provided detailed steps for installation and execution of the program to ensure the user can work with it smoothly. 
The updated README file should assist the user in quickly and effectively using the program.